Added .bat file for Node modules
You MUST run this .bat file at the root of the project for it to work properly. This file works with Windows, but I think Linux has a package you can install that will let you run .bat files. If not, let me know and I'll see if there's something I can do to help.